Seperate Select and MultiSelect component
